go back to classic extensions format
